+++ front/temporal/triggers/common/activities.ts
