+++ front/temporal/triggers/common/config.ts
