+++ front/temporal/triggers/common/worker.ts
